Deleted TruthTable File and added a more organised file structure
