+++ iengine.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31112.23
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TruthTable", "TruthTable\TruthTable.csproj", "{81A386A5-6D19-4A16-A12C-E07F369D28D2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "iengine", "iengine\iengine.csproj", "{589924F5-3693-4396-9B0C-DBA0A5491900}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{81A386A5-6D19-4A16-A12C-E07F369D28D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{81A386A5-6D19-4A16-A12C-E07F369D28D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{81A386A5-6D19-4A16-A12C-E07F369D28D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{81A386A5-6D19-4A16-A12C-E07F369D28D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{589924F5-3693-4396-9B0C-DBA0A5491900}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{589924F5-3693-4396-9B0C-DBA0A5491900}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{589924F5-3693-4396-9B0C-DBA0A5491900}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{589924F5-3693-4396-9B0C-DBA0A5491900}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CF607684-BAD3-421E-82BA-48B91674EB0C}
+		SolutionGuid = {1924A216-1671-467B-BDF2-7AAAC31395A6}
 	EndGlobalSection
 EndGlobal
